refactor: move DoubleEncodedJSONMiddleware to a separate module and simplify event ID generation in streaming service
