+++ src/agent_server/middleware/__init__.py
@@ -1,0 +1,3 @@
+from .double_encoded_json import DoubleEncodedJSONMiddleware
+
+__all__ = ["DoubleEncodedJSONMiddleware"]
