+++ src/agent_server/utils/__init__.py
@@ -1,0 +1,9 @@
+from .sse_utils import (
+    generate_event_id,
+    extract_event_sequence
+)
+
+__all__ = [
+    "generate_event_id",
+    "extract_event_sequence"
+]
